backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.30
google-genai
python-dotenv==1.0.1
Pillow==10.3.0
numpy==1.26.4
tensorflow==2.16.1
reportlab==4.2.2
qrcode[pil]==7.4.2
Werkzeug==3.0.3
gunicorn==22.0.0
email-validator==2.1.1
psycopg2-binary
requests==2.32.3
